api_transformer/requirements.txt
openvino == 2024.6.0
numpy <2.0.0
Pillow
fastapi
uvicorn
